requirements.txt
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
